requirements.txt
black==21.9b0
djangorestframework==3.12.4
factory-boy==3.2.0
isort==5.9.3
